Revert "Revert "Double to integer""
This reverts commit 90c15a616904c66604ab3b47f0374a320a5a4f0e.

+++ cult-of-tim-auth/src/main/java/com/cult_of_tim/auth/cultoftimauth/model/User.java
@@ -25,13 +25,13 @@ public class User {
     //@Column(name = "user_email")
     private String email;
 
-    private double balance;
+    private Integer balance;
 
-    public double getBalance() {
+    public Integer getBalance() {
         return balance;
     }
 
-    public void setBalance(double balance) {
+    public void setBalance(Integer balance) {
         this.balance = balance;
     }
